cluster gcn, finalize homogeneous
