--- examples/pytorch/__temporary__/dglnew/dataloading/__init__.py
@@ -1,2 +1,0 @@
-from .dataloader import *
-from .neighbor_sampler import *
