+++ examples/pytorch/__temporary__/dglnew/graph/other_feature.py
@@ -1,9 +1,7 @@
-import dgl
-from .graph import DGLGraphStorage
 
 # A GraphStorage class where ndata and edata can be any FeatureStorage but
-# otherwise the same as DGLGraph.
-class OtherFeatureGraphStorage(DGLGraphStorage):
+# otherwise the same as the wrapped DGLGraph.
+class OtherFeatureGraphStorage(object):
     def __init__(self, g, ndata=None, edata=None):
         self.g = g
         self._ndata = ndata or {}
@@ -16,3 +14,13 @@ class OtherFeatureGraphStorage(DGLGraphStorage):
     @property
     def edata(self):
         return self._edata
+
+    def __getattr__(self, key):
+        # I wrote it in this way because I'm too lazy to write "def sample_neighbors"
+        # or stuff like that.
+        if key in ['ntypes', 'etypes', 'canonical_etypes', 'sample_neighbors',
+                   'subgraph', 'edge_subgraph', 'find_edges', 'num_nodes']:
+            # Delegate to the wrapped DGLGraph instance.
+            return getattr(self.g, key)
+        else:
+            return super().__getattr__(key)
