--- examples/pytorch/__temporary__/dglnew/storages/__init__.py
@@ -1,3 +1,0 @@
-from .base import *
-from .tensor import *
-from .numpy import *
