+++ python/dgl/dataloading/asyncio_wrapper.py
@@ -1,3 +1,4 @@
+"""Python asyncio wrapper class."""
 import asyncio
 import threading
 # Solution to enable running asyncio.run in Jupyter notebooks.
