+++ python/dgl/sampling/neighbor.py
@@ -1,11 +1,15 @@
 """Neighbor sampling APIs"""
 
+from collections.abc import Mapping
+import numpy as np
+
 from .._ffi.function import _init_api
 from .. import backend as F
 from ..base import DGLError, EID
 from ..heterograph import DGLHeteroGraph
 from .. import ndarray as nd
-from .. import utils
+from .. import utils, transform
+from ..utils import recursive_apply, recursive_apply_pair
 
 __all__ = [
     'sample_etype_neighbors',
@@ -148,6 +152,73 @@ def sample_etype_neighbors(g, nodes, etype_field, fanout, edge_dir='in', prob=No
 
 DGLHeteroGraph.sample_etype_neighbors = utils.alias_func(sample_etype_neighbors)
 
+def _locate_eids_to_exclude(frontier_parent_eids, exclude_eids):
+    """Find the edges whose IDs in parent graph appeared in exclude_eids.
+
+    Note that both arguments are numpy arrays or numpy dicts.
+    """
+    func = lambda x, y: np.isin(x, y).nonzero()[0]
+    result = recursive_apply_pair(frontier_parent_eids, exclude_eids, func)
+    return recursive_apply(result, F.zerocopy_from_numpy)
+
+class _EidExcluder():
+    def __init__(self, exclude_eids):
+        device = None
+        if isinstance(exclude_eids, Mapping):
+            for _, v in exclude_eids.items():
+                if device is None:
+                    device = F.context(v)
+                    break
+        else:
+            device = F.context(exclude_eids)
+        self._exclude_eids = None
+        self._filter = None
+
+        if device == F.cpu():
+            # TODO(nv-dlasalle): Once Filter is implemented for the CPU, we
+            # should just use that irregardless of the device.
+            self._exclude_eids = (
+                recursive_apply(exclude_eids, F.zerocopy_to_numpy)
+                if exclude_eids is not None else None)
+        else:
+            self._filter = recursive_apply(exclude_eids, utils.Filter)
+
+    def _find_indices(self, parent_eids):
+        """ Find the set of edge indices to remove.
+        """
+        if self._exclude_eids is not None:
+            parent_eids_np = recursive_apply(parent_eids, F.zerocopy_to_numpy)
+            return _locate_eids_to_exclude(parent_eids_np, self._exclude_eids)
+        else:
+            assert self._filter is not None
+            func = lambda x, y: x.find_included_indices(y)
+            return recursive_apply_pair(self._filter, parent_eids, func)
+
+    def __call__(self, frontier):
+        parent_eids = frontier.edata[EID]
+        located_eids = self._find_indices(parent_eids)
+
+        if not isinstance(located_eids, Mapping):
+            # (BarclayII) If frontier already has a EID field and located_eids is empty,
+            # the returned graph will keep EID intact.  Otherwise, EID will change
+            # to the mapping from the new graph to the old frontier.
+            # So we need to test if located_eids is empty, and do the remapping ourselves.
+            if len(located_eids) > 0:
+                frontier = transform.remove_edges(
+                    frontier, located_eids, store_ids=True)
+                frontier.edata[EID] = F.gather_row(parent_eids, frontier.edata[EID])
+        else:
+            # (BarclayII) remove_edges only accepts removing one type of edges,
+            # so I need to keep track of the edge IDs left one by one.
+            new_eids = parent_eids.copy()
+            for k, v in located_eids.items():
+                if len(v) > 0:
+                    frontier = transform.remove_edges(
+                        frontier, v, etype=k, store_ids=True)
+                    new_eids[k] = F.gather_row(parent_eids[k], frontier.edges[k].data[EID])
+            frontier.edata[EID] = new_eids
+        return frontier
+
 def sample_neighbors(g, nodes, fanout, edge_dir='in', prob=None, replace=False,
                      copy_ndata=True, copy_edata=True, _dist_training=False,
                      exclude_edges=None, output_device=None):
@@ -221,8 +292,7 @@ def sample_neighbors(g, nodes, fanout, edge_dir='in', prob=None, replace=False,
     Returns
     -------
     DGLGraph
-        A sampled subgraph containing only the sampled neighboring edges, with the
-        same device as the input graph.
+        A sampled subgraph containing only the sampled neighboring edges.
 
     Notes
     -----
@@ -287,6 +357,21 @@ def sample_neighbors(g, nodes, fanout, edge_dir='in', prob=None, replace=False,
     tensor([False, False, False])
 
     """
+    if g.device == 'cpu':
+        frontier = _sample_neighbors(
+            g, nodes, fanout, edge_dir=edge_dir, prob=prob, replace=replace,
+            exclude_edges=exclude_edges)
+    else:
+        frontier = _sample_neighbors(
+            g, nodes, fanout, edge_dir=edge_dir, prob=prob, replace=replace)
+        if exclude_edges is not None:
+            eid_excluder = _EidExcluder(exclude_edges)
+            frontier = eid_excluder(frontier)
+    return frontier if output_device is None else frontier.to(output_device)
+
+def _sample_neighbors(g, nodes, fanout, edge_dir='in', prob=None, replace=False,
+                      copy_ndata=True, copy_edata=True, _dist_training=False,
+                      exclude_edges=None):
     if not isinstance(nodes, dict):
         if len(g.ntypes) > 1:
             raise DGLError("Must specify node type when the graph is not homogeneous.")
@@ -362,7 +447,7 @@ def sample_neighbors(g, nodes, fanout, edge_dir='in', prob=None, replace=False,
         for i, etype in enumerate(ret.canonical_etypes):
             ret.edges[etype].data[EID] = induced_edges[i]
 
-    return ret if output_device is None else ret.to(output_device)
+    return ret
 
 DGLHeteroGraph.sample_neighbors = utils.alias_func(sample_neighbors)
